--- Assignment/Assignment.Interfaces/Assignment.Interfaces.csproj
@@ -1,5 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <RootNamespace>Assignment</RootNamespace>
-  </PropertyGroup>
-</Project>
