--- Assignment/Assignment.sln
@@ -1,72 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 18
-VisualStudioVersion = 18.0.11201.2
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment", "Assignment\Assignment.csproj", "{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment.Interfaces", "Assignment.Interfaces\Assignment.Interfaces.csproj", "{6C4185C8-3398-49F8-B147-3B4418E2F922}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{E316FCD7-C0C8-4F2F-87FB-F92921756F02}"
-	ProjectSection(SolutionItems) = preProject
-		Directory.Build.props = Directory.Build.props
-		Directory.Build.targets = Directory.Build.targets
-		..\global.json = ..\global.json
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment.Tests", "Assignment.Tests\Assignment.Tests.csproj", "{0081689F-904E-4E2F-ADFF-EF85CE769A53}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Debug|x64.Build.0 = Debug|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Debug|x86.Build.0 = Debug|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Release|x64.ActiveCfg = Release|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Release|x64.Build.0 = Release|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Release|x86.ActiveCfg = Release|Any CPU
-		{F71D0FF8-246D-4D7E-9C1B-B6C15FE85EF8}.Release|x86.Build.0 = Release|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Debug|x64.Build.0 = Debug|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Debug|x86.Build.0 = Debug|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Release|x64.ActiveCfg = Release|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Release|x64.Build.0 = Release|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Release|x86.ActiveCfg = Release|Any CPU
-		{6C4185C8-3398-49F8-B147-3B4418E2F922}.Release|x86.Build.0 = Release|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Debug|x64.Build.0 = Debug|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Debug|x86.Build.0 = Debug|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Release|x64.ActiveCfg = Release|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Release|x64.Build.0 = Release|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Release|x86.ActiveCfg = Release|Any CPU
-		{0081689F-904E-4E2F-ADFF-EF85CE769A53}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {96A872D5-ACBB-46D6-B685-43C8B647A67C}
-	EndGlobalSection
-EndGlobal
